change scarf to have a single size
Set the default size for the scarf to `'xs'` - bit of a hack, as there is no
actual size for the scarf, but think it is a good compromise.
Ensure that when the scarf page loads it checks for the availability of the
product, so that we don't show the add to cart button in case our single size is
out of stock.