Implemented the `update-user-data` route with legacy handler
Also refactored a bit the `EmailStripperPipe`, move out the logic from it, so
that it can be used in the `UserDataDto` class as well. Moved the email trimming
into that new logic file.

Done the same on the user update's `UserDataDto` - meaning the trimming also
happens in the dto.

Changed the `get-user-data` route as well so that transformation happens in the
legacy handler and the new route returns a flat object.