Got the application to a runnable state
The database connection wasn't stable: due to TypeORM not being singleton, think
this is due to that we run the app via nextjs, that is keep re-building parts of
it.

With the `TypeOrmConfigService` we keep disconnect and then re-connect to the
database. To avoid getting the error, where it cannot find an entity that was
already initialized.

_Error:_
> EntityMetadataNotFound: No metadata for "<entity-name>" was found.

The `entity-name` was one of two: Address or Product.

This was needed so that we can run the whole thing in `dev` (`npm run
dev-experimental`), other way upon each page navigation we would crash the
back-end with the mentioned issue.