fix(i18n): correct the way we transfer for one language to another
Previously on the product detail page it wasn't possible to go to another
language as the `router.pathname` contained the path template (`/[idname]`)
instead of the actual path.

Didn't find a good solution on the net, logging the router noticed that the
`asPath` field contains the value that I would need. So hopefully it will
continue to work in the future as well.