refactor: moved the ping logic into a react hook
The ping component dealt with pinging the server and ensuring that the session
token is written to a cookie. Now this logic has no implications on the view,
but a background task that needs to be done and we need to ensure that a session
id is assigned at some point. As such moved the logic into a hook so that it can
be used from all the pages and it can update the session once we have it.

Session is really needed for the operation of the basket and also the checkout,
like this I'm not sure if a hook is good enough for getting the session as
without it we cannot operate, but don't have any better idea as of now.